// src/database/types/public/Users.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for users */
export type Id = string;

/** Represents the table public.users */
export default interface Users {
  id: Id;

  username: string;

  password_hash: string;

  created_at: Date | null;

  updated_at: Date | null;
}

/** Represents the initializer for the table public.users */
export interface UsersInitializer {
  id: Id;

  username: string;

  password_hash: string;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date | null;

  /** Default value: CURRENT_TIMESTAMP */
  updated_at?: Date | null;
}

/** Represents the mutator for the table public.users */
export interface UsersMutator {
  id?: Id;

  username?: string;

  password_hash?: string;

  created_at?: Date | null;

  updated_at?: Date | null;
}